requirements.txt
python-dotenv==1.0.0
openai==1.12.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.1
seaborn==0.12.2
jupyter==1.0.0
pytest==7.4.0
crewai==0.28.1
langchain==0.0.335
langchain-openai==0.0.5
pydantic==2.5.2 